Revert "Feature/check activities sw"
