+++ src/ducks/Signals/signalFormManager/signalCrudForm/formParts/channels/TriggerChannelSettings.js
@@ -5,37 +5,17 @@ import EmailSetting from '../../../../../../pages/Account/EmailSetting'
 import ConnectTelegramBlock from '../../../../../../pages/Account/ConnectTelegramBlock'
 import SettingsTelegramNotifications from '../../../../../../pages/Account/SettingsTelegramNotifications'
 import SettingsEmailNotifications from '../../../../../../pages/Account/SettingsEmailNotifications'
-import SettingsSonarWebPushNotifications, {
-  getSanSonarSW
-} from '../../../../../../pages/Account/SettingsSonarWebPushNotifications'
-import ShowIf from '../../../../../../components/ShowIf/ShowIf'
 import styles from './TriggerChannelSettings.module.scss'
 
-const TriggerChannelSettings = ({
-  isTelegramSettings,
-  isEmailSettings,
-  isBeta
-}) => {
+const TriggerChannelSettings = ({ isTelegramSettings, isEmailSettings }) => {
   const [open, setOpen] = useState(false)
-  const [isWebPushEnabled, setWebPushEnabled] = useState(true)
 
   useEffect(
     () => {
       !isTelegramSettings && !isEmailSettings && setOpen(false)
-
-      if (!isBeta) {
-        navigator.serviceWorker &&
-          navigator.serviceWorker.getRegistrations &&
-          navigator.serviceWorker.getRegistrations().then(registrations => {
-            const sw = getSanSonarSW(registrations)
-            setWebPushEnabled(!!sw)
-          })
-      }
     },
     [isEmailSettings, isTelegramSettings]
   )
-
-  const showTrigger = isTelegramSettings || isEmailSettings || !isWebPushEnabled
 
   return (
     <>
@@ -44,7 +24,9 @@ const TriggerChannelSettings = ({
         onClose={() => setOpen(false)}
         onOpen={() => setOpen(true)}
         trigger={
-          showTrigger && <span className={styles.connect}>Open settings</span>
+          (isTelegramSettings || isEmailSettings) && (
+            <span className={styles.connect}>Open settings</span>
+          )
         }
         title='Notification settings'
       >
@@ -62,12 +44,6 @@ const TriggerChannelSettings = ({
               <SettingsTelegramNotifications classes={styles} />
             </>
           )}
-
-          {!isWebPushEnabled && (
-            <ShowIf beta>
-              <SettingsSonarWebPushNotifications classes={styles} />
-            </ShowIf>
-          )}
         </Dialog.ScrollContent>
       </Dialog>
     </>
@@ -77,11 +53,9 @@ const TriggerChannelSettings = ({
 const mapStateToProps = ({
   user: {
     data: { email = '' }
-  },
-  rootUi: { isBetaModeEnabled }
+  }
 }) => ({
-  email,
-  isBeta: isBetaModeEnabled
+  email
 })
 
 const enhance = connect(mapStateToProps)
