+++ src/ducks/Signals/signalFormManager/signalCrudForm/formParts/channels/TriggerChannelSettings.module.scss
@@ -33,7 +33,7 @@
   margin-left: auto;
 }
 
-.inputContainer {
+.emailContainer {
   @include responsive('phone', 'phone-xs') {
     display: flex;
     flex-direction: column;
@@ -41,14 +41,14 @@
   }
 }
 
-.inputContainerRight {
+.emailContainerRight {
   @include responsive('phone', 'phone-xs') {
     display: flex;
     margin-top: 10px;
   }
 }
 
-.inputContainerLeft {
+.emailContainerLeft {
   @include responsive('phone', 'phone-xs') {
     margin: 0;
     width: 100%;
@@ -57,6 +57,6 @@
   }
 }
 
-.inputContainerLabel {
+.emailContainerLabel {
   margin: 0 0 0 auto;
 }
