+++ src/pages/Account/AccountPage.module.scss
@@ -15,10 +15,6 @@
 .tabs {
   width: 100%;
   margin-bottom: 25px;
-
-  @include responsive('phone', 'phone-xs') {
-    padding: 0 20px;
-  }
 }
 
 .tab {
@@ -69,13 +65,6 @@
 
   &_apikey {
     display: flex;
-    margin-left: auto;
-
-    @include responsive('phone', 'phone-xs') {
-      margin: 10px 0 0 0;
-      display: flex;
-      justify-content: center;
-    }
   }
 
   &__description {
@@ -98,6 +87,10 @@
     border: 1px solid var(--porcelain);
     display: flex;
     align-items: center;
+  }
+
+  :global(.container) {
+    margin-top: -16px !important;
   }
 
   a[href="https://www.bancor.network"] {
@@ -129,54 +122,20 @@
     &:last-child {
       margin: 0;
     }
-
-    @include responsive('phone', 'phone-xs') {
-      padding: 15px;
-    }
-  }
-}
-
-.inputLabels {
-  display: flex;
-  flex-direction: column;
-  word-break: break-word;
-
-  @include responsive('phone', 'phone-xs') {
-    flex-direction: column;
-  }
-}
-
-.inputBlock {
-  margin-right: 5px;
-  display: flex;
-
-  @include responsive('phone', 'phone-xs') {
-    width: 100%;
-    align-items: flex-start;
-    flex-direction: column !important;
   }
 }
 
 .form {
   position: relative;
-  display: flex;
-
-  @include responsive('phone', 'phone-xs') {
-    flex-direction: column;
-  }
 
   &__input {
+    position: absolute;
+    left: -9999px;
     margin-left: 16px;
-    display: none;
+    width: 320px;
 
     &_edit {
-      display: block;
-      width: 320px;
-
-      @include responsive('phone', 'phone-xs') {
-        margin: 0;
-        width: 100%;
-      }
+      position: static;
     }
   }
 
@@ -271,41 +230,9 @@
 }
 
 .metamask_connect {
-  margin-left: auto;
-
   @include responsive('phone', 'phone-xs') {
     margin: auto;
     margin-top: 10px;
-  }
-}
-
-.digest {
-  width: 100%;
-  display: flex;
-  align-items: center;
-
-  @include responsive('phone', 'phone-xs') {
-    display: flex;
-    flex-direction: column;
-  }
-}
-
-.neuro {
-  width: 100%;
-  display: flex;
-  align-items: center;
-
-  @include responsive('phone', 'phone-xs') {
-    display: flex;
-    flex-direction: column;
-  }
-}
-
-.digestSelector {
-  margin-left: auto;
-
-  @include responsive('phone', 'phone-xs') {
-    margin: 10px 0 0 0;
   }
 }
 
@@ -334,12 +261,12 @@
   }
 }
 
+.emailBlock {
+  margin-right: 5px;
+}
+
 .settingBlock {
   width: 100%;
   display: flex;
   flex-direction: row;
-}
-
-.pushLoader {
-  font-size: 5px;
 }
