+++ src/pages/Account/SettingsGeneral.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { connect } from 'react-redux'
 import { Toggle, Label } from '@santiment-network/ui'
 import Settings from './Settings'
-import UsernameSetting from './userName/UsernameSetting'
+import UsernameSetting from './UsernameSetting'
 import EmailSetting from './EmailSetting'
 import * as actions from '../../actions/types'
 import styles from './AccountPage.module.scss'
