--- src/pages/Account/userName/UsernameSettings.module.scss
@@ -1,25 +1,0 @@
-@import '~@santiment-network/ui/mixins.scss';
-
-.inputContainer {
-  @include responsive('phone', 'phone-xs') {
-    display: flex;
-    flex-direction: column;
-    height: auto !important;
-  }
-}
-
-.inputContainerRight {
-  @include responsive('phone', 'phone-xs') {
-    display: flex;
-    margin-top: 10px;
-  }
-}
-
-.inputContainerLeft {
-  @include responsive('phone', 'phone-xs') {
-    margin: 0;
-    width: 100%;
-    display: flex;
-    flex-direction: row;
-  }
-}
