+++ src/pages/SonarFeed/SonarFeedPage.js
@@ -134,7 +134,6 @@ const SonarFeed = ({
         <Switch>
           {isUserLoading && <PageLoader className={styles.loader} />}
           {!isUserLoading && !isLoggedIn ? <InsightUnAuthPage /> : ''}
-          <Route path={tabs[0].index} component={tabs[0].component} />
           {tabs.map(({ index, component }) => (
             <Route key={index} path={index} component={component} />
           ))}
