+++ src/serviceWorker.js
@@ -8,107 +8,74 @@ const isLocalhost = Boolean(
     )
 )
 
-const addSw = (swUrl, config) => {
-  if (isLocalhost) {
-    // This is running on localhost. Let's check if a service worker still exists or not.
-    checkValidServiceWorker(swUrl, config)
-
-    // Add some additional logging to localhost, pointing developers to the
-    // service worker/PWA documentation.
-    navigator.serviceWorker.ready.then(() => {
-      console.log('Sanbase is being served cache-first by a service')
-    })
-  } else {
-    // Is not localhost. Just register service worker
-    registerValidSW(swUrl, config)
-  }
-}
-
-const canExecuteSw = () => {
-  // The URL constructor is available in all browsers that support SW.
-  // Our service worker won't work if PUBLIC_URL is on a different origin
-  // from what our page is served on. This might happen if a CDN is used to
-  // serve assets; see https://github.com/facebook/create-react-app/issues/2374
-  const publicUrl = new URL(process.env.PUBLIC_URL, window.location.origin)
-  return publicUrl.origin === window.location.origin
-}
-
 export function register (config) {
-  if (
-    process.env.NODE_ENV === 'production' &&
-    'serviceWorker' in navigator &&
-    canExecuteSw()
-  ) {
-    addSw(`${process.env.PUBLIC_URL}/service-worker.js`, config)
-  }
-}
-
-export function registerSonarActivitiesSw (config) {
-  if ('serviceWorker' in navigator && canExecuteSw()) {
-    console.log('Register sonar service worker')
-    addSw(`${process.env.PUBLIC_URL}/san-sonar-service-worker.js`, config)
-  }
-}
-
-export const requestNotificationPermission = unRegisterCallback => {
-  const permission = window.Notification.requestPermission()
-  // value of permission can be 'granted', 'default', 'denied'
-  // granted: user has accepted the request
-  // default: user has dismissed the notification permission popup by clicking on x
-  // denied: user has denied the request.
-
-  permission.then(p => {
-    if (p !== 'granted') {
-      console.warn('Permission not granted for Notification')
-      unRegisterCallback && unRegisterCallback()
+  if (process.env.NODE_ENV === 'production' && 'serviceWorker' in navigator) {
+    // The URL constructor is available in all browsers that support SW.
+    const publicUrl = new URL(process.env.PUBLIC_URL, window.location.href)
+    if (publicUrl.origin !== window.location.origin) {
+      // Our service worker won't work if PUBLIC_URL is on a different origin
+      // from what our page is served on. This might happen if a CDN is used to
+      // serve assets; see https://github.com/facebook/create-react-app/issues/2374
+      return
     }
-  })
+
+    window.addEventListener('load', () => {
+      const swUrl = `${process.env.PUBLIC_URL}/service-worker.js`
+
+      if (isLocalhost) {
+        // This is running on localhost. Let's check if a service worker still exists or not.
+        checkValidServiceWorker(swUrl, config)
+
+        // Add some additional logging to localhost, pointing developers to the
+        // service worker/PWA documentation.
+        navigator.serviceWorker.ready.then(() => {
+          console.log('Sanbase is being served cache-first by a service')
+        })
+      } else {
+        // Is not localhost. Just register service worker
+        registerValidSW(swUrl, config)
+      }
+    })
+  }
 }
 
-function registerValidSW (
-  swUrl,
-  { callback, hideRegistrationChecking, onUpdate, onSuccess } = {}
-) {
+function registerValidSW (swUrl, config) {
   navigator.serviceWorker
     .register(swUrl)
     .then(registration => {
-      if (!hideRegistrationChecking) {
-        registration.onupdatefound = () => {
-          const installingWorker = registration.installing
-          if (installingWorker == null) {
-            return
-          }
-          installingWorker.onstatechange = () => {
-            if (installingWorker.state === 'installed') {
-              if (navigator.serviceWorker.controller) {
-                // At this point, the updated precached content has been fetched,
-                // but the previous service worker will still serve the older
-                // content until all client tabs are closed.
-                console.log(
-                  'New content is available and will be used when all ' +
-                    'tabs for this page are closed. See https://bit.ly/CRA-PWA.'
-                )
+      registration.onupdatefound = () => {
+        const installingWorker = registration.installing
+        if (installingWorker == null) {
+          return
+        }
+        installingWorker.onstatechange = () => {
+          if (installingWorker.state === 'installed') {
+            if (navigator.serviceWorker.controller) {
+              // At this point, the updated precached content has been fetched,
+              // but the previous service worker will still serve the older
+              // content until all client tabs are closed.
+              console.log(
+                'New content is available and will be used when all ' +
+                  'tabs for this page are closed. See https://bit.ly/CRA-PWA.'
+              )
 
-                // Execute callback
-                if (onUpdate) {
-                  onUpdate(registration)
-                }
-              } else {
-                // At this point, everything has been precached.
-                // It's the perfect time to display a
-                // "Content is cached for offline use." message.
-                console.log('Content is cached for offline use.')
+              // Execute callback
+              if (config && config.onUpdate) {
+                config.onUpdate(registration)
+              }
+            } else {
+              // At this point, everything has been precached.
+              // It's the perfect time to display a
+              // "Content is cached for offline use." message.
+              console.log('Content is cached for offline use.')
 
-                // Execute callback
-                if (onSuccess) {
-                  onSuccess(registration)
-                }
+              // Execute callback
+              if (config && config.onSuccess) {
+                config.onSuccess(registration)
               }
             }
           }
         }
-      } else {
-        callback && callback()
       }
     })
     .catch(error => {
